LPS-199140 Move class SharingEntryDropdownItemContributorRegistry to class ViewSharedAssetsDisplayContext, and apply to usages
